Add support for changing base images (#27)
Projects can now define `UBUNTU_VERSION` to any version of Ubuntu that is desired (supporting both numerical versions like `16.04` and named aliases like `focal`). This will be used as the Ubuntu version for compilers and for the runtime container.

Internally, there's a big change to the various Docker images and containers managed by PwnableHarness. Here's the current overview:

* `pwnmake-v<pwnmake version>`: This is an Ubuntu 24.04 image used to run `make` commands, and is invoked by the `pwnmake` CLI script.
* `pwncc-<ubuntu version>-v<pwnableharness-version>`: These are a collection of tiny Ubuntu images with compilers installed. Any compiler or linker invocations used to build challenges are now run through these `pwncc` containers, to easily allow compiling for specific Ubuntu versions.
* `base-<ubuntu version>-v<pwnableharness-version>`: These are a collection of Ubuntu images that serve as the base runtime environment for PwnableHarness challenges. They contain `libpwnableharness*.so` and `pwnableserver` and have the environment set up, ready to serve a challenge.

From a usage perspective, the only change is the new `UBUNTU_VERSION` variable. Compilation commands will be somewhat slower due to now running each one in a separate Docker container, but that's an acceptable trade-off since PwnableHarness projects are typically small.